scripts/requirements-aca.txt
# Optional: only for scripts/aca_apply_multicontainer.py (Azure CLI YAML patch)
PyYAML>=6.0.1
